+++ include/IceCraft/world.h
@@ -18,7 +18,7 @@ enum WorldModificationType
 struct WorldModification
 {
     enum WorldModificationType type;
-    float x, y, z;
+    int x, y, z;
     unsigned material_id;  // Doesn't have to be initialized when type == DESTROYED
 };
 
@@ -39,7 +39,7 @@ void generate_lobby_world(struct World *world, struct TextureAtlas *texture_atla
 
 void generate_ice_world(struct World *world, struct TextureAtlas *texture_atlas);
 
-void world_place_block(struct World *world, float x, float y, float z, unsigned texture_id, struct TextureAtlas *texture_atlas);
-void world_destroy_block_at_position(struct World *world, float x, float y, float z);
+void world_place_block(struct World *world, int x, int y, int z, unsigned texture_id, struct TextureAtlas *texture_atlas);
+void world_destroy_block_at_position(struct World *world, int x, int y, int z);
 
 void world_free_chunk(struct World *world);
